+++ source/Irrlicht/CIrrDeviceLinux.h
@@ -216,11 +216,11 @@ namespace irr
 
 				if (CursorPos.X < 0)
 					CursorPos.X = 0;
-				if (CursorPos.X > Device->Width)
+				if (CursorPos.X > (s32) Device->Width)
 					CursorPos.X = Device->Width;
 				if (CursorPos.Y < 0)
 					CursorPos.Y = 0;
-				if (CursorPos.Y > Device->Height)
+				if (CursorPos.Y > (s32) Device->Height)
 					CursorPos.Y = Device->Height;
 #endif
 			}
